Complete the clickLemonImage() function and rewrite the setViewElements() func
Instead of changing the textView and the ImageView in the when condition in the setViewElements() func call, the code now uses 2 int variables to store the id's of the corresponding strings and images. Then at the end the Views are updated to the corresponding values of the resouce ID's just 1 time. The clickLemonImage() func has been implemented according to the conditions given